+++ packages/observation-extractor/lib/extractor/ergo/ergoExtractor.ts
@@ -20,13 +20,13 @@ export class ErgoObservationExtractor extends AbstractExtractor<Transaction> {
 
   constructor(
     dataSource: DataSource,
-    tokens: RosenTokens,
+    tokens: TokenMap,
     address: string,
     logger?: AbstractLogger
   ) {
     super();
     this.dataSource = dataSource;
-    this.tokens = new TokenMap(tokens);
+    this.tokens = tokens;
     this.logger = logger ? logger : new DummyLogger();
     this.actions = new ObservationEntityAction(dataSource, this.logger);
     this.extractor = new ErgoNodeRosenExtractor(address, tokens, this.logger);