+++ packages/watcher-data-extractor/tests/extractor/utilsVariable.mock.ts
@@ -34,7 +34,7 @@ export const sampleEventEntity: EventTriggerEntity = {
   eventId: 'eventId',
   extractor: 'extractorId',
   boxId: 'id',
-  boxSerialized: 'boxSerialized',
+  serialized: 'boxSerialized',
   block: 'hash',
   height: 10,
   toAddress: 'cardanoAddr2',
@@ -1122,7 +1122,4 @@ export const spendTriggerTx: Transaction = {
   ],
 };
 
-export const testTokenMap = {
-  idKeys: {},
-  tokens: [],
-};
+export const testTokenMap = [];