+++ packages/address-extractor/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @rosen-bridge/address-extractor
 
+## 6.0.1
+
+### Patch Changes
+
+- Update node version to 20
+- Update logger package
+- Updated dependencies
+  - @rosen-bridge/abstract-extractor@1.0.1
+
+## 6.0.0
+
+### Major Changes
+
+- Update database entity and action to extend the abstract class
+
+### Minor Changes
+
+- Update db action interface to support callbacks
+
+### Patch Changes
+
+- Updated dependencies
+  - @rosen-bridge/abstract-extractor@1.0.0
+
 ## 5.0.8
 
 ### Patch Changes