+++ packages/fraud-extractor/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @rosen-bridge/fraud-extractor
 
+## 2.0.12
+
+### Patch Changes
+
+- Update node version to 20
+- Update logger package
+- Updated dependencies
+  - @rosen-bridge/abstract-extractor@1.0.1
+  - @rosen-bridge/scanner@5.0.2
+
+## 2.0.11
+
+### Patch Changes
+
+- Updated dependencies
+  - @rosen-bridge/abstract-extractor@1.0.0
+  - @rosen-bridge/scanner@5.0.1
+
 ## 2.0.10
 
 ### Patch Changes