+++ packages/watcher-data-extractor/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @rosen-bridge/watcher-data-extractor
 
+## 11.0.0
+
+### Major Changes
+
+- Update tokens package version to 2.0.0
+
+## 10.0.1
+
+### Patch Changes
+
+- Update node version to 20
+- Updated dependencies
+  - @rosen-bridge/abstract-extractor@1.0.1
+  - @rosen-bridge/scanner@5.0.2
+
+## 10.0.0
+
+### Major Changes
+
+- Update database entity and action to extend the abstract class
+
+### Minor Changes
+
+- Update db action interface to support callbacks
+
+### Patch Changes
+
+- Optimize `hasData` function not to deserialize all unrelated boxes
+- Updated dependencies
+  - @rosen-bridge/abstract-extractor@1.0.0
+  - @rosen-bridge/scanner@5.0.1
+
 ## 9.0.0
 
 ### Major Changes